Fix multi-line variable definitions (again)
The last fix didn't really work. This is a lot more code, but it does
actually work. The problem is that the spec requires the "begin" and
"end" must not overlap, but the following simple rules result in ambi-
guity:

> "begin": "(\\w+)\\s*(\\+?=)"
> "end": "[^+=]\s*$"

The ambiguity occurs with this code:

> my_var =
>     "value"

"Begin" matches "my_var =" and "end" matches " =$". The implemented so-
lution splits the single rule for variable assignments into two: one for
the single-line declaration ending with "= value" and one for the multi-
line declaration ending with "=".

Lists and maps on variables are still broken and will be fixed
separately.